Split ``sphinx.ext.intersphinx._load``
